Move types related to api responses to api.types.ts
