+++ ui/catalog/src/components/DataSourceConnectorsTable/types.ts
@@ -1,4 +1,5 @@
 import type { DataTableHeader } from "@carbon/react";
+import type { ConnectorStatus } from "@/types/api.types";
 import type {
   BaseTableState,
   SharedTableAction,
@@ -8,33 +9,6 @@ import {
   isSharedTableAction,
   setLoading,
 } from "@/components/Table/utils/reducerUtils";
-
-export type ConnectorStatus = "Connected" | "Offline";
-
-export interface ConnectorProvider {
-  id: string;
-  name: string;
-}
-
-export interface DataSourceConnectorApiResponse {
-  id: string;
-  name: string;
-  type: string;
-  provider: ConnectorProvider;
-  status: ConnectorStatus;
-  message: string;
-  connected_services: number;
-  metadata: Record<string, unknown>;
-  created_at: string;
-  updated_at: string;
-}
-
-export interface DataSourceConnectorsListResponse {
-  data: DataSourceConnectorApiResponse[];
-  total: number;
-  page: number;
-  page_size: number;
-}
 
 export interface DataSourceConnectorRow {
   id: string;
